==> scripts/data_preprocessing/00_extract_cleaned_data.ipynb <==
# Import relevant libraries

import os
import tarfile

# Extract cleaned_data.tar.gz

# set base_path
base_path = os.path.join('../..')

# set file path
file_path = os.path.join(base_path, 'datasets', 'cleaned_data.tar.gz')
destination = os.path.join(base_path, 'datasets')

# extract .tar.gz
tar = tarfile.open(file_path, "r:gz")
tar.extractall(path = destination)
tar.close()
print(f'File is extracted to {destination}')
