==> scripts/data_preprocessing/01_Select Features.ipynb <==
# Import relevant libraries

import os
import pandas as pd

# Setup

# disable false positive warnings
pd.options.mode.chained_assignment = None

# set base_path
base_path = os.path.join('..', '..')

file_path = os.path.join(base_path, 'datasets', 'cleaned_data.csv')
df = pd.read_csv(file_path)

# Preprocessing

df.info()

# select columns that are necessary
df1 = df[['CompanyName', 'CompanyNumber', 'CompanyCategory', 'CompanyStatus', 'DissolutionDate', 'IncorporationDate', 'SICCode.SicText_1', 'SICCode.SicText_2', 'SICCode.SicText_3', 'SICCode.SicText_4']]

df1['IncorporationDate'] = pd.to_datetime(df1['IncorporationDate'])

df1['year'] = df1['IncorporationDate'].dt.year
df1['month'] = df1['IncorporationDate'].dt.month

df1 = df1.sort_values('IncorporationDate', ascending=True)

# save file
file_path = os.path.join(base_path, 'datasets', 'features_data.csv')
df1.to_csv(file_path, index = False)
