==> scripts/data_preprocessing/02_final_data_cleaning.ipynb <==
# Import relevant libraries

import os
import pandas as pd
import numpy as np

# Load Data & Merge

# set file path
base_path = '../..'
file_path_1 = os.path.join(base_path, 'datasets',
                           'tech_roundabout_coordinates.csv')

# load data
tech_roundabout_full = pd.read_csv(file_path_1)

# set file path
file_path_2 = os.path.join(base_path, 'datasets', 'features_data.csv')

# load data
features = pd.read_csv(file_path_2)

# merge both data
all_data = pd.merge(tech_roundabout_full,
                    features,
                    on='CompanyName',
                    how='inner')

# Split Incorporation Date to find Sector Codes and Values

# sort value by date
all_data = all_data.sort_values('IncorporationDate', ascending=True)

# remove all companies before 1990
all_data = all_data[all_data['year'] > 1990]

# split sector code so we can identify sector values

x = all_data['SICCode.SicText_1'].str.rsplit(pat='-', expand=True)

y = all_data['SICCode.SicText_2'].str.rsplit(pat='-', expand=True)

z = all_data['SICCode.SicText_3'].str.rsplit(pat='-', expand=True)

q = all_data['SICCode.SicText_4'].str.rsplit(pat='-', expand=True)

all_data['sector_code_1'] = x[0]
all_data['sector_name_1'] = x[1]
all_data['sector_code_2'] = y[0]
all_data['sector_name_2'] = y[1]
all_data['sector_code_3'] = z[0]
all_data['sector_name_3'] = z[1]
all_data['sector_code_4'] = q[0]
all_data['sector_name_4'] = q[1]

all_data.info()

# Clean Data and remove unnecessary columns

# remove spaces in columns
col_names = [col_name.strip() for col_name in all_data.columns]

# set column names
all_data.columns = col_names

# remove columns
columns_to_remove = [
    'RegAddress.PostCode', 'RegAddress.PostTown', 'SICCode.SicText_1',
    'SICCode.SicText_2', 'SICCode.SicText_3', 'SICCode.SicText_4'
]

cleaned_data = all_data.drop(columns_to_remove, axis=1)

cleaned_data.loc[cleaned_data['CompanyStatus'] != 'Active']

# Create dictionary that shows count of each sector code

# find unqiue sector values
unique_codes = np.unique(cleaned_data['sector_code_1'], return_counts=True)

# create dictionary with count of each sector values

d = {}
for idx, val in enumerate(unique_codes[0]):
    d[val] = unique_codes[1][idx]

sorted(d.items(), key=lambda item: item[1])

# Create new df specific to sectors

# find names for certain sector values

# create new data frame for sector values
columns = [
    'IncorporationDate',
    'year',
    'month',
    'CompanyName',
    'CompanyNumber',
    'CompanyStatus',
    'sector_code_1',
    'sector_name_1',
]

# create new df with columns
sector_data = cleaned_data[columns]

# sort by sector code
sector_data.sort_values(by=['sector_code_1', 'sector_name_1'],
                        ascending=True).head(20)

# save data as csv
save_path = os.path.join(base_path, 'datasets', 'final_cleaned_data.csv')
sector_data.to_csv(save_path, index=False)
